requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.6
django-allauth==0.63.3
django-cleanup==8.1.0
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
msgpack==1.0.8
multidict==6.0.5
phonenumbers==8.13.40
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
sqlparse==0.5.0
tomli==2.0.1
twilio==9.2.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
zope.interface==7.0.2
